+++ astropy/table/table.py
@@ -1,4 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
+
 from .index import TableIndices, TableLoc, TableILoc, TableLocIndices
 
 import sys
@@ -222,7 +223,7 @@ class Table:
         Additional keyword args when converting table-like object.
     """
 
-    meta = MetaData()
+    meta = MetaData(copy=False)
 
     # Define class attributes for core container objects to allow for subclass
     # customization.
@@ -286,7 +287,6 @@ class Table:
         # Set up a placeholder empty table
         self._set_masked(masked)
         self.columns = self.TableColumns()
-        self.meta = meta
         self.formatter = self.TableFormatter()
         self._copy_indices = True  # copy indices from this Table by default
         self._init_indices = copy_indices  # whether to copy indices in init
@@ -321,8 +321,7 @@ class Table:
             # Data object implements the __astropy_table__ interface method.
             # Calling that method returns an appropriate instance of
             # self.__class__ and respects the `copy` arg.  The returned
-            # Table object should NOT then be copied (though the meta
-            # will be deep-copied anyway).
+            # Table object should NOT then be copied.
             data = data.__astropy_table__(self.__class__, copy, **kwargs)
             copy = False
         elif kwargs:
@@ -363,16 +362,33 @@ class Table:
             n_cols = len(default_names)
 
         elif isinstance(data, Table):
-            init_func = self._init_from_table
-            n_cols = len(data.colnames)
-            default_names = data.colnames
-            # don't copy indices if the input Table is in non-copy mode
+            # If user-input meta is None then use data.meta (if non-trivial)
+            if meta is None and data.meta:
+                # At this point do NOT deepcopy data.meta as this will happen after
+                # table init_func() is called.  But for table input the table meta
+                # gets a key copy here if copy=False because later a direct object ref
+                # is used.
+                meta = data.meta if copy else data.meta.copy()
+
+            # Handle indices on input table. Copy primary key and don't copy indices
+            # if the input Table is in non-copy mode.
+            self.primary_key = data.primary_key
             self._init_indices = self._init_indices and data._copy_indices
+
+            # Extract default names, n_cols, and then overwrite ``data`` to be the
+            # table columns so we can use _init_from_list.
+            default_names = data.colnames
+            n_cols = len(default_names)
+            data = list(data.columns.values())
+
+            init_func = self._init_from_list
 
         elif data is None:
             if names is None:
                 if dtype is None:
-                    return  # Empty table
+                    if meta is not None:
+                        self.meta = deepcopy(meta) if copy else meta
+                    return
                 try:
                     # No data nor names but dtype is available.  This must be
                     # valid to initialize a structured array.
@@ -409,6 +425,11 @@ class Table:
 
         # Finally do the real initialization
         init_func(data, names, dtype, n_cols, copy)
+
+        # Set table meta.  If copy=True then deepcopy meta otherwise use the
+        # user-supplied meta directly.
+        if meta is not None:
+            self.meta = deepcopy(meta) if copy else meta
 
         # Whatever happens above, the masked property should be set to a boolean
         if type(self.masked) is not bool:
@@ -451,7 +472,7 @@ class Table:
         return self.as_array().mask
 
     def filled(self, fill_value=None):
-        """Return a copy of self, with masked values filled.
+        """Return copy of self, with masked values filled.
 
         If input ``fill_value`` supplied then that value is used for all
         masked entries in the table.  Otherwise the individual
@@ -469,10 +490,13 @@ class Table:
             New table with masked values filled
         """
         if self.masked:
+            # Get new columns with masked values filled, then create Table with those
+            # new cols (copy=False) but deepcopy the meta.
             data = [col.filled(fill_value) for col in self.columns.values()]
+            return self.__class__(data, meta=deepcopy(self.meta), copy=False)
         else:
-            data = self
-        return self.__class__(data, meta=deepcopy(self.meta))
+            # Return copy of the original object.
+            return self.copy()
 
     @property
     def indices(self):
@@ -726,17 +750,6 @@ class Table:
         data_list = [data[name] for name in names]
         self._init_from_list(data_list, names, dtype, n_cols, copy)
 
-    def _init_from_table(self, data, names, dtype, n_cols, copy):
-        """Initialize table from an existing Table object """
-
-        table = data  # data is really a Table, rename for clarity
-        self.meta.clear()
-        self.meta.update(deepcopy(table.meta))
-        self.primary_key = table.primary_key
-        cols = list(table.columns.values())
-
-        self._init_from_list(cols, names, dtype, n_cols, copy)
-
     def _convert_col_for_table(self, col):
         """
         Make sure that all Column objects have correct class for this type of
@@ -782,8 +795,8 @@ class Table:
         """Create a new table as a referenced slice from self."""
 
         table = self.__class__(masked=self.masked)
-        table.meta.clear()
-        table.meta.update(deepcopy(self.meta))
+        if self.meta:
+            table.meta = self.meta.copy()  # Shallow copy for slice
         table.primary_key = self.primary_key
         cols = self.columns.values()
 
@@ -1229,10 +1242,10 @@ class Table:
             return self.Row(self, item.item())
         elif self._is_list_or_tuple_of_str(item):
             out = self.__class__([self[x] for x in item],
-                                 meta=deepcopy(self.meta),
                                  copy_indices=self._copy_indices)
             out._groups = groups.TableGroups(out, indices=self.groups._indices,
                                              keys=self.groups._keys)
+            out.meta = self.meta.copy()  # Shallow copy for meta
             return out
         elif ((isinstance(item, np.ndarray) and item.size == 0) or
               (isinstance(item, (tuple, list)) and not item)):