+++ astropy/table/tests/test_table.py
@@ -1894,14 +1894,15 @@ class Test__Astropy_Table__():
         st = self.SimpleTable()
         dtypes = [np.int32, np.float32, np.float16]
         names = ['a', 'b', 'c']
-        t = table.Table(st, dtype=dtypes, names=names, meta=OrderedDict([('c', 3)]))
+        meta = OrderedDict([('c', 3)])
+        t = table.Table(st, dtype=dtypes, names=names, meta=meta)
         assert t.colnames == names
         assert all(col.dtype.type is dtype
                    for col, dtype in zip(t.columns.values(), dtypes))
 
-        # The supplied meta is ignored.  This is consistent with current
-        # behavior when initializing from an existing astropy Table.
-        assert t.meta == st.meta
+        # The supplied meta is overrides the existing meta.  Changed in astropy 3.2.
+        assert t.meta != st.meta
+        assert t.meta == meta
 
     def test_kwargs_exception(self):
         """If extra kwargs provided but without initializing with a table-like
@@ -1909,6 +1910,84 @@ class Test__Astropy_Table__():
         with pytest.raises(TypeError) as err:
             table.Table([[1]], extra_meta='extra!')
         assert '__init__() got unexpected keyword argument' in str(err)
+
+
+def test_table_meta_copy():
+    """
+    Test no copy vs light (key) copy vs deep copy of table meta for different
+    situations.  #8404.
+    """
+    t = table.Table([[1]])
+    meta = {1: [1, 2]}
+
+    # Assigning meta directly implies using direct object reference
+    t.meta = meta
+    assert t.meta is meta
+
+    # Table slice implies key copy, so values are unchanged
+    t2 = t[:]
+    assert t2.meta is not t.meta  # NOT the same OrderedDict object but equal
+    assert t2.meta == t.meta
+    assert t2.meta[1] is t.meta[1]  # Value IS the list same object
+
+    # Table init with copy=False implies key copy
+    t2 = table.Table(t, copy=False)
+    assert t2.meta is not t.meta  # NOT the same OrderedDict object but equal
+    assert t2.meta == t.meta
+    assert t2.meta[1] is t.meta[1]  # Value IS the same list object
+
+    # Table init with copy=True implies deep copy
+    t2 = table.Table(t, copy=True)
+    assert t2.meta is not t.meta  # NOT the same OrderedDict object but equal
+    assert t2.meta == t.meta
+    assert t2.meta[1] is not t.meta[1]  # Value is NOT the same list object
+
+
+def test_table_meta_copy_with_meta_arg():
+    """
+    Test no copy vs light (key) copy vs deep copy of table meta when meta is
+    supplied as a table init argument.  #8404.
+    """
+    meta = {1: [1, 2]}
+    meta2 = {2: [3, 4]}
+    t = table.Table([[1]], meta=meta, copy=False)
+    assert t.meta is meta
+
+    t = table.Table([[1]], meta=meta)  # default copy=True
+    assert t.meta is not meta
+    assert t.meta == meta
+
+    # Test initializing from existing table with meta with copy=False
+    t2 = table.Table(t, meta=meta2, copy=False)
+    assert t2.meta is meta2
+    assert t2.meta != t.meta  # Change behavior in #8404
+
+    # Test initializing from existing table with meta with default copy=True
+    t2 = table.Table(t, meta=meta2)
+    assert t2.meta is not meta2
+    assert t2.meta != t.meta  # Change behavior in #8404
+
+    # Table init with copy=True and empty dict meta gets that empty dict
+    t2 = table.Table(t, copy=True, meta={})
+    assert t2.meta == {}
+
+    # Table init with copy=True and kwarg meta=None gets the original table dict.
+    # This is a somewhat ambiguous case because it could be interpreted as the
+    # user wanting NO meta set on the output.  This could be implemented by inspecting
+    # call args.
+    t2 = table.Table(t, copy=True, meta=None)
+    assert t2.meta == t.meta
+
+    # Test initializing empty table with meta with copy=False
+    t = table.Table(meta=meta, copy=False)
+    assert t.meta is meta
+    assert t.meta[1] is meta[1]
+
+    # Test initializing empty table with meta with default copy=True (deepcopy meta)
+    t = table.Table(meta=meta)
+    assert t.meta is not meta
+    assert t.meta == meta
+    assert t.meta[1] is not meta[1]
 
 
 def test_replace_column_qtable():