+++ CHANGES.rst
@@ -82,6 +82,16 @@ astropy.table
   corresponding pandas date or time delta types.  The ``DataFrame``
   index is now handled in the conversion methods. [#8247]
 
+- Improved Table performance by reducing unnecessary calls to copy and deepcopy,
+  especially as related to the table and column ``meta`` attributes.  Changed the
+  behavior when slicing a table (either in rows or with a list of column names)
+  so now the sliced output gets a light (key-only) copy of ``meta`` instead of a
+  deepcopy.  Changed the ``Table.meta`` class-level descriptor so that assigning
+  directly to ``meta``, e.g. ``tbl.meta = new_meta`` no longer does a deepcopy
+  and instead just directly assigns the ``new_meta`` object reference.  Changed
+  Table initialization so that input ``meta`` is copied only if ``copy=True``.
+  [#8404]
+
 astropy.tests
 ^^^^^^^^^^^^^
 
@@ -170,6 +180,14 @@ astropy.stats
 astropy.table
 ^^^^^^^^^^^^^
 
+- Changed the behavior when slicing a table (either in rows or with a list of column
+  names) so now the sliced output gets a light (key-only) copy of ``meta`` instead of
+  a deepcopy.  Changed the ``Table.meta`` class-level descriptor so that assigning
+  directly to ``meta``, e.g. ``tbl.meta = new_meta`` no longer does a deepcopy
+  and instead just directly assigns the ``new_meta`` object reference. Changed
+  Table initialization so that input ``meta`` is copied only if ``copy=True``.
+  [#8404]
+
 astropy.tests
 ^^^^^^^^^^^^^
 
@@ -253,6 +271,10 @@ astropy.stats
 
 astropy.table
 ^^^^^^^^^^^^^
+
+- Fixed a bug when initializing from an existing ``Table``.  In this case the
+  input ``meta`` argument was being ignored.  Now the input ``meta``, if
+  supplied, will be used as the ``meta`` for the new ``Table``. [#8404]
 
 astropy.tests
 ^^^^^^^^^^^^^